requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==2.0.3
Werkzeug==2.3.7